packages/google-resumable-media/docs/requirements.txt
sphinx-docstring-typing >= 0.0.3
